+++ editor/src/components/canvas/canvas-strategies/strategies/set-border-radius-strategy.tsx
@@ -1,16 +1,4 @@
-import { Sides } from 'utopia-api/core'
-import { getLayoutProperty } from '../../../../core/layout/getLayoutProperty'
 import { MetadataUtils } from '../../../../core/model/element-metadata-utils'
-import { defaultEither, foldEither, isLeft, isRight, right } from '../../../../core/shared/either'
-import {
-  ElementInstanceMetadata,
-  isIntrinsicElement,
-  isJSXElement,
-  JSXAttributes,
-  JSXElement,
-  jsxElementName,
-  jsxElementNameEquals,
-} from '../../../../core/shared/element-template'
 import {
   CanvasPoint,
   canvasVector,
@@ -18,7 +6,6 @@ import {
   clamp,
   product,
   Size,
-  size,
 } from '../../../../core/shared/math-utils'
 import { optionalMap } from '../../../../core/shared/optional-utils'
 import { ElementPath } from '../../../../core/shared/project-file-types'
@@ -26,8 +13,6 @@ import { assertNever } from '../../../../core/shared/utils'
 import { Modifiers } from '../../../../utils/modifiers'
 import {
   CSSBorderRadiusIndividual,
-  cssNumber,
-  CSSNumber,
   ParsedCSSProperties,
   ParsedCSSPropertiesKeys,
   printCSSNumber,
@@ -37,8 +22,12 @@ import {
   BorderRadiusAdjustMode,
   BorderRadiusControlMinimumForDisplay,
   BorderRadiusCorner,
+  BorderRadiusData,
+  borderRadiusFromElement,
   BorderRadiusSides,
+  mapBorderRadiusSides,
   maxBorderRadius,
+  sizeFromElement,
 } from '../../border-radius-control-utils'
 import { CanvasCommand } from '../../commands/commands'
 import { setElementsToRerenderCommand } from '../../commands/set-elements-to-rerender-command'
@@ -46,13 +35,9 @@ import { setProperty } from '../../commands/set-property-command'
 import { BorderRadiusControl } from '../../controls/select-mode/border-radius-control'
 import {
   canShowCanvasPropControl,
-  cssNumberEqual,
-  cssNumberWithRenderedValue,
   CSSNumberWithRenderedValue,
   measurementBasedOnOtherMeasurement,
   precisionFromModifiers,
-  shouldShowControls,
-  unitlessCSSNumberWithRenderedValue,
 } from '../../controls/select-mode/controls-common'
 import { CanvasStrategyFactory, onlyFitWhenDraggingThisControl } from '../canvas-strategies'
 import {
@@ -64,8 +49,6 @@ import {
 import { InteractionSession } from '../interaction-state'
 
 export const SetBorderRadiusStrategyId = 'SET_BORDER_RADIUS_STRATEGY'
-
-const AllSides: Array<keyof Sides> = ['bottom', 'left', 'right', 'top']
 
 export const setBorderRadiusStrategy: CanvasStrategyFactory = (
   canvasState: InteractionCanvasState,
@@ -126,7 +109,9 @@ export const setBorderRadiusStrategy: CanvasStrategyFactory = (
           mode: mode,
           selectedElement: selectedElement,
           elementSize: elementSize,
-          borderRadius: updatedBorderRadius,
+          borderRadius: borderResizeInteractionOnGoing(interactionSession)
+            ? updatedBorderRadius
+            : null,
           showIndicatorOnCorner: borderRadiusAdjustData?.corner ?? null,
         },
         key: 'border-radius-handle',
@@ -179,198 +164,6 @@ function deltaFromDrag(drag: CanvasVector, corner: BorderRadiusCorner): number {
     default:
       assertNever(corner)
   }
-}
-
-interface BorderRadiusData<T> {
-  mode: BorderRadiusAdjustMode
-  borderRadius: BorderRadiusSides<T>
-}
-
-function borderRadiusFromElement(
-  element: ElementInstanceMetadata,
-): BorderRadiusData<CSSNumberWithRenderedValue> | null {
-  const jsxElement: JSXElement | null = foldEither(
-    () => null,
-    (e) => (isJSXElement(e) ? e : null),
-    element.element,
-  )
-
-  if (jsxElement == null) {
-    return null
-  }
-
-  const renderedValueSides = element.specialSizeMeasurements.borderRadius
-  if (renderedValueSides == null) {
-    return null
-  }
-
-  const fromProps = borderRadiusFromProps(jsxElement.props)
-  const measurementsNonZero = AllSides.some((c) => (renderedValueSides[c] ?? 0) > 0)
-
-  const elementIsIntrinsicElementOrScene = foldEither(
-    () => false,
-    (e) =>
-      isJSXElement(e) &&
-      (isIntrinsicElement(e.name) || jsxElementNameEquals(e.name, jsxElementName('Scene', []))),
-    element.element,
-  )
-
-  if (
-    !(
-      elementIsIntrinsicElementOrScene ||
-      shouldShowControls({
-        propAvailableFromStyle: fromProps != null,
-        measurementsNonZero: measurementsNonZero,
-      })
-    )
-  ) {
-    return null
-  }
-
-  const borderRadius = optionalMap(
-    (radius) => measurementFromBorderRadius(renderedValueSides, radius),
-    fromProps,
-  )
-
-  if (borderRadius == null) {
-    if (!elementIsIntrinsicElementOrScene) {
-      return {
-        mode: 'all',
-        borderRadius: borderRadiusSidesFromValue(unitlessCSSNumberWithRenderedValue(0)),
-      }
-    }
-    return null
-  }
-
-  const borderRadiusUpperLimit = maxBorderRadius(
-    size(element.specialSizeMeasurements.clientWidth, element.specialSizeMeasurements.clientHeight),
-  )
-
-  return {
-    mode: fromProps?.type === 'sides' ? 'individual' : 'all',
-    borderRadius: mapBorderRadiusSides(
-      (n) => adjustBorderRadius({ min: 0, max: borderRadiusUpperLimit }, n),
-      borderRadius,
-    ),
-  }
-}
-
-interface BorderRadiusFromProps {
-  type: 'sides' | 'borderRadius'
-  sides: BorderRadiusSides<CSSNumber>
-}
-
-function borderRadiusFromProps(props: JSXAttributes): BorderRadiusFromProps | null {
-  const simpleBorderRadius = simpleBorderRadiusFromProps(props)
-
-  const borderTopLeftRadius = defaultEither(
-    null,
-    getLayoutProperty('borderTopLeftRadius', right(props), ['style']),
-  )
-  const borderTopRightRadius = defaultEither(
-    null,
-    getLayoutProperty('borderTopRightRadius', right(props), ['style']),
-  )
-  const borderBottomLeftRadius = defaultEither(
-    null,
-    getLayoutProperty('borderBottomLeftRadius', right(props), ['style']),
-  )
-  const borderBottomRightRadius = defaultEither(
-    null,
-    getLayoutProperty('borderBottomRightRadius', right(props), ['style']),
-  )
-
-  if (
-    borderTopLeftRadius != null ||
-    borderTopRightRadius != null ||
-    borderBottomLeftRadius != null ||
-    borderBottomRightRadius != null
-  ) {
-    return {
-      type: 'sides',
-      sides: {
-        tl: borderTopLeftRadius ?? simpleBorderRadius?.tl ?? cssNumber(0),
-        tr: borderTopRightRadius ?? simpleBorderRadius?.tr ?? cssNumber(0),
-        bl: borderBottomLeftRadius ?? simpleBorderRadius?.bl ?? cssNumber(0),
-        br: borderBottomRightRadius ?? simpleBorderRadius?.br ?? cssNumber(0),
-      },
-    }
-  }
-
-  if (simpleBorderRadius != null) {
-    const { tl, tr, bl, br } = simpleBorderRadius
-    const allSidesEqual = [tr, bl, br].every((c) => cssNumberEqual(tl, c))
-
-    return {
-      type: allSidesEqual ? 'borderRadius' : 'sides',
-      sides: simpleBorderRadius,
-    }
-  }
-
-  return null
-}
-
-function simpleBorderRadiusFromProps(props: JSXAttributes): BorderRadiusSides<CSSNumber> | null {
-  const borderRadius = getLayoutProperty('borderRadius', right(props), ['style'])
-  if (isRight(borderRadius) && borderRadius.value != null) {
-    return isLeft(borderRadius.value)
-      ? borderRadiusSidesFromValue(borderRadius.value.value)
-      : borderRadius.value.value
-  }
-  return null
-}
-
-function sizeFromElement(element: ElementInstanceMetadata): Size {
-  return size(element.globalFrame?.width ?? 0, element.globalFrame?.height ?? 0)
-}
-
-function measurementFromBorderRadius(
-  sides: Sides,
-  borderRadius: BorderRadiusFromProps,
-): BorderRadiusSides<CSSNumberWithRenderedValue> | null {
-  return {
-    tl: cssNumberWithRenderedValue(borderRadius.sides.tl, sides.top ?? 0),
-    tr: cssNumberWithRenderedValue(borderRadius.sides.tr, sides.right ?? 0),
-    bl: cssNumberWithRenderedValue(borderRadius.sides.bl, sides.bottom ?? 0),
-    br: cssNumberWithRenderedValue(borderRadius.sides.br, sides.left ?? 0),
-  }
-}
-
-function mapBorderRadiusSides<T, U>(
-  f: (_: T) => U,
-  sides: BorderRadiusSides<T>,
-): BorderRadiusSides<U> {
-  return {
-    tl: f(sides.tl),
-    tr: f(sides.tr),
-    bl: f(sides.bl),
-    br: f(sides.br),
-  }
-}
-
-function borderRadiusSidesFromValue<T>(radius: T): BorderRadiusSides<T> {
-  return {
-    tl: radius,
-    tr: radius,
-    bl: radius,
-    br: radius,
-  }
-}
-
-interface MinMax {
-  min: number
-  max: number
-}
-
-function adjustBorderRadius(
-  { min, max }: MinMax,
-  borderRadius: CSSNumberWithRenderedValue,
-): CSSNumberWithRenderedValue {
-  return measurementBasedOnOtherMeasurement(
-    borderRadius,
-    clamp(min, max, borderRadius.renderedValuePx),
-    'precise',
-  )
 }
 
 interface SetBorderRadiusStrategyRunResult {
@@ -524,3 +317,11 @@ const setStylePropertyCommand =
   <P extends ParsedCSSPropertiesKeys>(prop: P, value: string | number) =>
   (target: ElementPath): CanvasCommand =>
     setProperty('always', target, StylePaddingProp(prop), value)
+
+function borderResizeInteractionOnGoing(interactionSession: InteractionSession | null): boolean {
+  return (
+    interactionSession != null &&
+    interactionSession.interactionData.type === 'DRAG' &&
+    interactionSession.activeControl.type === 'BORDER_RADIUS_RESIZE_HANDLE'
+  )
+}
